+++ src/Files.App/Converters/Converters.cs
@@ -11,7 +11,7 @@ namespace Files.App.Converters
 	/// </summary>
 	/// <typeparam name="TSource">The source type.</typeparam>
 	/// <typeparam name="TTarget">The target type.</typeparam>
-	public abstract class ValueConverter<TSource, TTarget> : IValueConverter
+	internal abstract class ValueConverter<TSource, TTarget> : IValueConverter
 	{
 		/// <summary>
 		/// Converts a source value to the target type.
@@ -89,7 +89,7 @@ namespace Files.App.Converters
 	/// <summary>
 	/// The base class for converting instances of type T to object and vice versa.
 	/// </summary>
-	public abstract class ToObjectConverter<T> : ValueConverter<T?, object?>
+	internal abstract class ToObjectConverter<T> : ValueConverter<T?, object?>
 	{
 		/// <summary>
 		/// Converts a source value to the target type.
@@ -119,7 +119,7 @@ namespace Files.App.Converters
 	/// <summary>
 	/// Converts a boolean to and from a visibility value.
 	/// </summary>
-	public class InverseBooleanConverter : ValueConverter<bool, bool>
+	internal class InverseBooleanConverter : ValueConverter<bool, bool>
 	{
 		/// <summary>
 		/// Converts a source value to the target type.
@@ -146,7 +146,7 @@ namespace Files.App.Converters
 		}
 	}
 
-	public class NullToTrueConverter : ValueConverter<object?, bool>
+	internal class NullToTrueConverter : ValueConverter<object?, bool>
 	{
 		/// <summary>
 		/// Determines whether an inverse conversion should take place.
@@ -179,7 +179,7 @@ namespace Files.App.Converters
 		}
 	}
 
-	public class StringNullOrWhiteSpaceToTrueConverter : ValueConverter<string, bool>
+	internal class StringNullOrWhiteSpaceToTrueConverter : ValueConverter<string, bool>
 	{
 		/// <summary>
 		/// Determines whether an inverse conversion should take place.
