+++ src/Files.App/Converters/DoubleArrayToStringConverter.cs
@@ -6,16 +6,17 @@ using System.Text;
 
 namespace Files.App.Converters
 {
-	internal class StringArrayToString : IValueConverter
+	internal class DoubleArrayToStringConverter : IValueConverter
 	{
 		public object Convert(object value, Type targetType, object parameter, string language)
 		{
-			var array = value as string[];
+			var array = value as double[];
 
-			if (array is null || !(array is string[]))
+			if (array is null)
 				return string.Empty;
 
 			var str = new StringBuilder();
+
 			foreach (var i in array)
 			{
 				str.Append(string.Format("{0}; ", i));
@@ -26,7 +27,20 @@ namespace Files.App.Converters
 
 		public object ConvertBack(object value, Type targetType, object parameter, string language)
 		{
-			return (value as string).Split("; ");
+			var strArray = (value as string).Split("; ");
+			var array = new double[strArray.Length];
+
+			for (int i = 0; i < strArray.Length; i++)
+			{
+				try
+				{
+					array[i] = double.Parse(strArray[i]);
+				}
+				catch (Exception)
+				{
+				}
+			}
+			return array;
 		}
 	}
 }
