+++ src/Files.App/Converters/ImageModelToImageConverter.cs
@@ -1,20 +1,19 @@
-﻿// Copyright (c) 2023 Files Community
+// Copyright (c) 2023 Files Community
 // Licensed under the MIT License. See the LICENSE.
 
-using CommunityToolkit.WinUI.Helpers;
+using Files.App.AppModels;
 using Microsoft.UI.Xaml.Data;
-using Microsoft.UI.Xaml.Media;
 
-namespace Files.App.ValueConverters
+namespace Files.App.Converters
 {
-	internal sealed class StringToBrushConverter : IValueConverter
+	internal sealed class ImageModelToImageConverter : IValueConverter
 	{
 		public object? Convert(object value, Type targetType, object parameter, string language)
 		{
-			if (value is not string strValue)
-				return null;
+			if (value is BitmapImageModel bitmapImageModel)
+				return bitmapImageModel.Image;
 
-			return new SolidColorBrush(ColorHelper.ToColor(strValue));
+			return null;
 		}
 
 		public object ConvertBack(object value, Type targetType, object parameter, string language)
