+++ src/Files.App/Converters/StringArrayToStringConverter.cs
@@ -6,17 +6,16 @@ using System.Text;
 
 namespace Files.App.Converters
 {
-	internal class DoubleArrayToString : IValueConverter
+	internal class StringArrayToStringConverter : IValueConverter
 	{
 		public object Convert(object value, Type targetType, object parameter, string language)
 		{
-			var array = value as double[];
+			var array = value as string[];
 
-			if (array is null)
+			if (array is null || !(array is string[]))
 				return string.Empty;
 
 			var str = new StringBuilder();
-
 			foreach (var i in array)
 			{
 				str.Append(string.Format("{0}; ", i));
@@ -27,20 +26,7 @@ namespace Files.App.Converters
 
 		public object ConvertBack(object value, Type targetType, object parameter, string language)
 		{
-			var strArray = (value as string).Split("; ");
-			var array = new double[strArray.Length];
-
-			for (int i = 0; i < strArray.Length; i++)
-			{
-				try
-				{
-					array[i] = double.Parse(strArray[i]);
-				}
-				catch (Exception)
-				{
-				}
-			}
-			return array;
+			return (value as string).Split("; ");
 		}
 	}
 }
