+++ src/Files.App/Converters/UInt32ToStringConverter.cs
@@ -5,7 +5,7 @@ using Microsoft.UI.Xaml.Data;
 
 namespace Files.App.Converters
 {
-	internal class UInt32ToString : IValueConverter
+	internal class UInt32ToStringConverter : IValueConverter
 	{
 		public object Convert(object value, Type targetType, object parameter, string language)
 		{
