+++ src/Files.App/ResourceDictionaries/PropertiesStyles.xaml
@@ -64,10 +64,10 @@
 		<Setter Property="FontSize" Value="14" />
 	</Style>
 
-	<converters:StringArrayToString x:Key="StringArrayToString" />
-	<converters:UInt32ToString x:Key="UInt32ToString" />
-	<converters:DoubleToString x:Key="DoubleToString" />
-	<converters:DateTimeOffsetToString x:Key="DateTimeOffsetToString" />
+	<converters:StringArrayToStringConverter x:Key="StringArrayToString" />
+	<converters:UInt32ToStringConverter x:Key="UInt32ToString" />
+	<converters:DoubleToStringConverter x:Key="DoubleToString" />
+	<converters:DateTimeOffsetToStringConverter x:Key="DateTimeOffsetToString" />
 
 	<Style x:Key="Separator" TargetType="MenuFlyoutSeparator">
 		<Setter Property="Grid.Column" Value="0" />
