+++ src/Files.App/Views/LayoutModes/ColumnViewBase.xaml
@@ -11,7 +11,6 @@
 	xmlns:local="using:Files.App"
 	xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
 	xmlns:uc="using:Files.App.UserControls"
-	xmlns:valueconverters="using:Files.App.ValueConverters"
 	xmlns:wct="using:CommunityToolkit.WinUI.UI"
 	xmlns:wctconverters="using:CommunityToolkit.WinUI.UI.Converters"
 	x:Name="PageRoot"
@@ -19,7 +18,7 @@
 
 	<local:StandardViewBase.Resources>
 
-		<valueconverters:StringToBrushConverter x:Key="StringToBrushConverter" />
+		<converters:StringToBrushConverter x:Key="StringToBrushConverter" />
 		<wctconverters:BoolNegationConverter x:Key="BoolNegationConverter" />
 		<wctconverters:BoolToVisibilityConverter
 			x:Key="NegatedBoolToVisibilityConverter"
@@ -33,7 +32,7 @@
 			x:Key="EmptyObjectToObjectConverter"
 			EmptyValue="Collapsed"
 			NotEmptyValue="Visible" />
-		<converters:BoolToSelectionMode x:Key="BoolToSelectionModeConverter" />
+		<converters:BoolToSelectionModeConverter x:Key="BoolToSelectionModeConverter" />
 
 		<Style TargetType="ListViewHeaderItem">
 			<Setter Property="FontFamily" Value="{ThemeResource ContentControlThemeFontFamily}" />
